Frontend assets get copied into the static bundle, static files use the copy.
The "build" dir from the theme and the "app-assets" dir from the app gets
copied to the newly created static styleguide bundle, and the link href's and
script src's of the static .html files point to these local-to-the-bundle
copies.

Not so nice: the aforementioned directories get copied over by shell cp command
via shell_exec() instead of some recursive php copy()'ing. Should be fixed.